pyproject.toml
[project]
name = "lab01-part2-18135122"
version = "1.0.0"
description = "NYC Taxi Data Engineering Pipeline - Containerized with Docker, Great Expectations, and BI"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=2.0",
    "psycopg2-binary>=2.9",
    "pyarrow>=15.0",
    "pyspark>=4.0",
    "sqlalchemy>=2.0",
    "great-expectations>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "requests>=2.31",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
